[3.8] Check whether process is used in ImportConfiguration before deletion
