+++ Kitodo/src/main/java/org/kitodo/production/forms/ProcessForm.java
@@ -94,7 +94,6 @@ public class ProcessForm extends TemplateBaseForm {
 
     private String processEditReferer = DEFAULT_LINK;
     private String taskEditReferer = DEFAULT_LINK;
-    private String errorMessage = "";
 
     private List<SelectItem> customColumns;
 
@@ -1094,20 +1093,6 @@ public class ProcessForm extends TemplateBaseForm {
     }
 
     /**
-     * Rename media files of all selected processes.
-     */
-    public void renameMedia() {
-        List<Process> processes = getSelectedProcesses();
-        errorMessage = ServiceManager.getFileService().tooManyProcessesSelectedForMediaRenaming(processes.size());
-        if (StringUtils.isBlank(errorMessage)) {
-            PrimeFaces.current().executeScript("PF('renameMediaConfirmDialog').show();");
-        } else {
-            Ajax.update("errorDialog");
-            PrimeFaces.current().executeScript("PF('errorDialog').show();");
-        }
-    }
-
-    /**
      * Start renaming media files of selected processes.
      */
     public void startRenaming() {
@@ -1125,14 +1110,6 @@ public class ProcessForm extends TemplateBaseForm {
     public String getMediaRenamingConfirmMessage() {
         return Helper.getTranslation("renameMediaForProcessesConfirmMessage",
                 String.valueOf(getSelectedProcesses().size()));
-    }
-
-    /**
-     * Get error message.
-     * @return error message
-     */
-    public String getErrorMessage() {
-        return errorMessage;
     }
 
     /**
