+++ Kitodo/src/main/java/org/kitodo/production/forms/ProcessListBaseView.java
@@ -22,12 +22,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.kitodo.api.dataformat.PhysicalDivision;
 import org.kitodo.api.dataformat.Workpiece;
 import org.kitodo.config.ConfigCore;
 import org.kitodo.config.enums.ParameterCore;
+import org.kitodo.data.database.beans.ImportConfiguration;
 import org.kitodo.data.database.beans.Process;
 import org.kitodo.data.database.exceptions.DAOException;
 import org.kitodo.data.exceptions.DataException;
@@ -42,6 +44,7 @@ import org.kitodo.production.process.ProcessMetadataStatistic;
 import org.kitodo.production.services.ServiceManager;
 import org.kitodo.production.services.data.ProcessService;
 import org.kitodo.production.services.dataformat.MetsService;
+import org.omnifaces.util.Ajax;
 import org.primefaces.PrimeFaces;
 import org.primefaces.component.datatable.DataTable;
 import org.primefaces.event.data.PageEvent;
@@ -49,6 +52,8 @@ import org.primefaces.model.charts.hbar.HorizontalBarChartModel;
 import org.primefaces.model.charts.pie.PieChartModel;
 
 public class ProcessListBaseView extends BaseForm {
+
+    private String errorMessage = "";
 
     private static final Logger logger = LogManager.getLogger(ProcessListBaseView.class);
     private ChartMode chartMode;
@@ -74,6 +79,14 @@ public class ProcessListBaseView extends BaseForm {
     public ProcessListBaseView() {
         super();
         super.setLazyDTOModel(new LazyProcessDTOModel(ServiceManager.getProcessService()));
+    }
+
+    /**
+     * Get error message.
+     * @return error message
+     */
+    public String getErrorMessage() {
+        return errorMessage;
     }
 
     /**
@@ -549,17 +562,28 @@ public class ProcessListBaseView extends BaseForm {
     public void delete(ProcessDTO processDTO) {
         try {
             Process process = ServiceManager.getProcessService().getById(processDTO.getId());
-            if (process.getChildren().isEmpty()) {
-                try {
-                    ProcessService.deleteProcess(process);
-                } catch (DataException | IOException e) {
-                    Helper.setErrorMessage(ERROR_DELETING, new Object[] {ObjectType.PROCESS.getTranslationSingular() },
-                            logger, e);
+            try {
+                List<ImportConfiguration> configurations = ServiceManager.getImportConfigurationService()
+                        .getProcessTemplateConfigurationByProcessId(process.getId());
+                if (!configurations.isEmpty()) {
+                    errorMessage = Helper.getTranslation("processUsedAsTemplateProcess", configurations.get(0).getTitle());
+                    Ajax.update("errorDialog");
+                    PrimeFaces.current().executeScript("PF('errorDialog').show();");
+                } else if (process.getChildren().isEmpty()) {
+                    try {
+                        ProcessService.deleteProcess(process.getId());
+                    } catch (DAOException | IOException | DataException e) {
+                        Helper.setErrorMessage(ERROR_DELETING, new Object[]{ObjectType.PROCESS.getTranslationSingular()}, logger, e);
+                    }
                 }
-            } else {
-                this.deleteProcessDialog = new DeleteProcessDialog();
-                this.deleteProcessDialog.setProcess(process);
-                PrimeFaces.current().executeScript("PF('deleteChildrenDialog').show();");
+                else {
+                    this.deleteProcessDialog = new DeleteProcessDialog();
+                    this.deleteProcessDialog.setProcess(process);
+                    PrimeFaces.current().executeScript("PF('deleteChildrenDialog').show();");
+                }
+            } catch (DAOException e) {
+                Helper.setErrorMessage(ERROR_LOADING_MANY, new Object[]{ObjectType.IMPORT_CONFIGURATION.getTranslationSingular()},
+                        logger, e);
             }
         } catch (DAOException e) {
             Helper.setErrorMessage(ERROR_DELETING, new Object[] {ObjectType.PROCESS.getTranslationSingular() }, logger,
@@ -591,6 +615,20 @@ public class ProcessListBaseView extends BaseForm {
         } catch (IOException | DAOException e) {
             Helper.setErrorMessage(e);
             return false;
+        }
+    }
+
+    /**
+     * Rename media files of all selected processes.
+     */
+    public void renameMedia() {
+        List<Process> processes = getSelectedProcesses();
+        errorMessage = ServiceManager.getFileService().tooManyProcessesSelectedForMediaRenaming(processes.size());
+        if (StringUtils.isBlank(errorMessage)) {
+            PrimeFaces.current().executeScript("PF('renameMediaConfirmDialog').show();");
+        } else {
+            Ajax.update("errorDialog");
+            PrimeFaces.current().executeScript("PF('errorDialog').show();");
         }
     }
