+++ Kitodo/src/main/java/org/kitodo/production/services/data/ImportConfigurationService.java
@@ -142,6 +142,18 @@ public class ImportConfigurationService extends SearchDatabaseService<ImportConf
     }
 
     /**
+     * Load and return all ImportConfiguration of type PROCESS_TEMPLATE that are configured with the given process.
+     * @param processId of the process to check
+     * @return List of ImportConfiguration  configured with the given process
+     * @throws DAOException when ImportConfigurations could not be loaded
+     */
+    public List<ImportConfiguration> getProcessTemplateConfigurationByProcessId(int processId) throws DAOException {
+        return getAllImportConfigurations(ImportConfigurationType.PROCESS_TEMPLATE).stream()
+                .filter(importConfiguration -> importConfiguration.getDefaultTemplateProcess().getId() == processId)
+                .collect(Collectors.toList());
+    }
+
+    /**
      * Load and return all ImportConfigurations sorted by title.
      * @return list of all ImportConfigurations sorted by title
      * @throws DAOException when ImportConfigurations could not be loaded
