+++ Kitodo/src/main/resources/messages/errors_de.properties
@@ -99,7 +99,7 @@ kitodoScript.importProcesses.indir.isNull=Parameter ''indir'' fehlt!
 kitodoScript.importProcesses.indir.isNoDirectory=''indir'' ist kein existierendes Verzeichnis!
 kitodoScript.importProcesses.indir.cannotExecute=Der Inhalt des Importverzeichnisses kann nicht abgerufen werden (fehlende Berechtigung \u201Cexecute\u201D)
 kitodoScript.importProcesses.missingChildren=Vorgang {0} enth\u004E Verweise auf fehlendene Kindvorg\u004Enge {1}
-kitodoScript.importProcesses.noProcessTitleRule=Regelsatz gibt keine Bildungsvorschrift für den ''processTitle'' für Division {0} an!
+kitodoScript.importProcesses.noProcessTitleRule=Regelsatz gibt keine Bildungsvorschrift f\u00FCr den ''processTitle'' f\u00FCr Division {0} an!
 kitodoScript.importProcesses.project.isNull=Parameter ''project'' fehlt!
 kitodoScript.importProcesses.project.isNoProjectID=''project'' ist keine g\u00FCltige Projekt-ID
 kitodoScript.importProcesses.project.noProjectWithID=Ein solches Projekt existiert nicht!
@@ -171,6 +171,7 @@ processTitleAlreadyInUse=Der Prozesstitel ''{0}'' wird bereits verwendet.
 processTitleEmpty=Kein Vorgangstitel angegeben.
 processTitleInvalid=Ung\u00FCltiger Titel f\u00FCr einen Vorgang. Er sollte mit dem regul\u00E4ren Ausdruck ''{0}'' \u00FCbereinstimmen.
 projectTitleAlreadyInUse=Der Projekttitel wird bereits verwendet.
+processUsedAsTemplateProcess=Vorgang kann nicht gel\u00F6scht werden, da er in mindestens einer Importkonfiguration als Vorlagenvorgang verwendet wird: ''{0}''
 propertiesForProcessNotSaved=Eigenschaften des Prozesses {0} konnten nicht gespeichert werden.
 
 # R
